backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
python-dotenv==1.0.0
notion-client==2.2.1
aiohttp==3.9.1
asyncio==3.4.3
schedule==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
Flask==3.0.0

